# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BARQ Agentic Incident Resolution Platform",
    "description": "Autonomous incident triage, diagnosis, and remediation platform with HITL governance.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/executions/{execution_id}": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "Get execution details by ID",
        "description": "Fetch the operational status and lifecycle attributes of a specific execution.",
        "operationId": "get_execution_api_v1_executions__execution_id__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Execution Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/executions/{execution_id}/trace": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "Get execution trace and node progression",
        "description": "Fetch the ordered sequence of workflow node states executed for this run.",
        "operationId": "get_execution_trace_api_v1_executions__execution_id__trace_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Execution Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{sys_id}/executions": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "List all executions for a ServiceNow incident",
        "description": "Fetch all execution audit runs associated with a 32-hex incident sys_id.",
        "operationId": "list_incident_executions_api_v1_incidents__sys_id__executions_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sys_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "description": "ServiceNow incident 32-character sys_id",
              "title": "Sys Id"
            },
            "description": "ServiceNow incident 32-character sys_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentExecutionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecutionNodeStateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Primary key of workflow_state record"
          },
          "execution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Execution Id",
            "description": "Foreign key to executions.execution_id"
          },
          "sequence_number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence Number",
            "description": "Sequence order of the node attempt (>= 1)"
          },
          "node_name": {
            "type": "string",
            "maxLength": 100,
            "title": "Node Name",
            "description": "Name of the workflow node"
          },
          "attempt": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Attempt",
            "description": "Attempt count for this specific node (>= 1)",
            "default": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "started",
              "succeeded",
              "failed",
              "blocked",
              "awaiting_approval",
              "skipped"
            ],
            "title": "Status",
            "description": "Node state status matching workflow_state.status constraint"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At",
            "description": "Timestamp when node attempt started"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At",
            "description": "Timestamp when node attempt concluded"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence",
            "description": "Evidence array gathered during this node execution"
          },
          "decision": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision",
            "description": "Node decision payload if applicable"
          },
          "state_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Snapshot",
            "description": "State snapshot object if captured"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "execution_id",
          "sequence_number",
          "node_name",
          "status",
          "started_at"
        ],
        "title": "ExecutionNodeStateResponse",
        "description": "Schema representing one node attempt from the 'workflow_state' table."
      },
      "ExecutionResponse": {
        "properties": {
          "execution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Execution Id",
            "description": "Primary key from executions.execution_id"
          },
          "event_record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Record Id",
            "description": "Foreign key to events.id"
          },
          "incident_sys_id": {
            "type": "string",
            "maxLength": 32,
            "title": "Incident Sys Id",
            "description": "ServiceNow incident 32-character sys_id"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "queued",
              "running",
              "awaiting_approval",
              "succeeded",
              "failed",
              "blocked",
              "abandoned"
            ],
            "title": "Status",
            "description": "Execution lifecycle status matching executions.status constraint"
          },
          "node_reached": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Reached",
            "description": "Latest workflow node entered"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name",
            "description": "LLM model identifier used for this execution"
          },
          "agent_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version",
            "description": "Version string of the executing agent"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At",
            "description": "Timestamp when execution started"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At",
            "description": "Timestamp when execution reached a terminal state"
          },
          "termination_cause": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Termination Cause",
            "description": "Explanation when execution reached terminal status"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Timestamp when execution record was last modified"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "execution_id",
          "event_record_id",
          "incident_sys_id",
          "status",
          "started_at",
          "updated_at"
        ],
        "title": "ExecutionResponse",
        "description": "Schema representing an execution record from the 'executions' table."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentExecutionsResponse": {
        "properties": {
          "incident_sys_id": {
            "type": "string",
            "maxLength": 32,
            "title": "Incident Sys Id",
            "description": "ServiceNow incident 32-character sys_id"
          },
          "executions": {
            "items": {
              "$ref": "#/components/schemas/ExecutionResponse"
            },
            "type": "array",
            "title": "Executions",
            "description": "List of executions associated with this incident"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "incident_sys_id"
        ],
        "title": "IncidentExecutionsResponse",
        "description": "Collection of execution records for a specific ServiceNow incident."
      },
...
      "TraceResponse": {
        "properties": {
          "execution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Execution Id",
            "description": "Unique execution identifier"
          },
          "incident_sys_id": {
            "type": "string",
            "maxLength": 32,
            "title": "Incident Sys Id",
            "description": "ServiceNow incident sys_id"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "queued",
              "running",
              "awaiting_approval",
              "succeeded",
              "failed",
              "blocked",
              "abandoned"
            ],
            "title": "Status",
            "description": "Current execution status"
          },
          "node_states": {
            "items": {
              "$ref": "#/components/schemas/ExecutionNodeStateResponse"
            },
            "type": "array",
            "title": "Node States",
            "description": "Ordered list of historical workflow node attempts from workflow_state"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "execution_id",
          "incident_sys_id",
          "status"
        ],
        "title": "TraceResponse",
        "description": "Execution trace progression combining execution status and workflow node states."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Bearer token authentication. Enter your token (without 'Bearer ' prefix).",
        "scheme": "bearer"
      }
    }
  }
}